+++ MahjongTools/Visualization.py
@@ -23,6 +23,7 @@
 # 胡牌(001)，弃牌(002, 003, 004)，暗杠(005)，补牌(006)
 
 import numpy as np
+import os
 
 TILE_NAMES = [
         ["一万", "两万", "三万", "四万", "五万", "六万", "七万", "八万", "九万"],
@@ -34,8 +35,25 @@ TILE_NAMES = [
 class EngineVisualization:
     """游戏过程可视化类，负责将当前局面在命令行画出来"""
     def __init__(self):
-        pass
+        self.true_tile_names = TILE_NAMES
     
+    def reset(self, god_id: int) -> None:
+        """重置可视化状态"""
+        self.true_tile_names = TILE_NAMES
+        god_tile_name = TILE_NAMES[god_id//9][god_id%9]
+        self.true_tile_names[god_id//9][god_id%9] = "白板"
+        self.true_tile_names[3][6] = god_tile_name # 将财神牌显示为白板，原本的财神牌信息放在白板位置
+        return
+
+    def draw_game(self, game_state: dict) -> None:
+        """绘制当前游戏状态，包括玩家手牌、牌墙剩余牌数、当前玩家等信息
+        Args:
+            game_state: 当前游戏状态对象，从MahjongEngine的get_visible_state方法获取"""
+        os.system('cls' if os.name == 'nt' else 'clear') # 每次绘制前先清屏，保持界面整洁
+        self.display_game_state(game_state)
+        self.display_last_action(game_state['last_action'])
+        return
+
     def display_game_state(self, game_state: dict) -> None:
         """显示当前游戏状态，包括玩家手牌、牌墙剩余牌数、当前玩家等信息
         Args:
@@ -47,34 +65,23 @@ class EngineVisualization:
         print("当前庄家玩家: 玩家", game_state['banker_index'],"，当前底分:", game_state['basic_score'])
         print("牌墙剩余牌数: ", int(game_state['remaining_tiles'] * 56), "，当前财神牌: ", TILE_NAMES[game_state['god_id']//9][game_state['god_id']%9])
         print("--------------------")
-        print("对家: 玩家", game_state['real_order'][2])
-        print("碰牌: ", player_meld[1]['peng'])
-        print("吃牌: ", player_meld[1]['chi'])
-        print("明杠: ", player_meld[1]['ming_gang'])
-        print("暗杠数量: ", player_meld[1]['an_gang_count'])
-        print("弃牌: ", player_meld[1]['discards'])
-        print("--------------------")
-        print("上家: 玩家", game_state['real_order'][3])
-        print("碰牌: ", player_meld[2]['peng'])
-        print("吃牌: ", player_meld[2]['chi'])
-        print("明杠: ", player_meld[2]['ming_gang'])
-        print("暗杠数量: ", player_meld[2]['an_gang_count'])
-        print("弃牌: ", player_meld[2]['discards'])
-        print("--------------------")
-        print("下家: 玩家", game_state['real_order'][1])
-        print("碰牌: ", player_meld[0]['peng'])
-        print("吃牌: ", player_meld[0]['chi'])
-        print("明杠: ", player_meld[0]['ming_gang'])
-        print("暗杠数量: ", player_meld[0]['an_gang_count'])
-        print("弃牌: ", player_meld[0]['discards'])
-        print("--------------------")
+        print_name_map = {0: "下", 1: "对", 2: "上"}
+        for i in range(3):
+            j = (i + 1) % 3
+            print(f"{print_name_map[j]}家: 玩家", game_state['real_order'][j+1])
+            if player_meld[j]['peng']:                  print("碰牌: ", player_meld[j]['peng'])
+            if player_meld[j]['chi']:                   print("吃牌: ", player_meld[j]['chi'])
+            if player_meld[j]['ming_gang']:             print("明杠: ", player_meld[j]['ming_gang'])
+            if player_meld[j]['an_gang_count'] > 0:     print("暗杠数量: ", player_meld[j]['an_gang_count'])
+            if player_meld[j]['discards']:              print("弃牌: ", player_meld[j]['discards'])
+            print("--------------------")
         print("当前玩家: 玩家", game_state['real_order'][0])
         print("手牌: ", current_player_meld['hand'])
-        print("碰牌: ", current_player_meld['peng'])
-        print("吃牌: ", current_player_meld['chi'])
-        print("明杠: ", current_player_meld['ming_gang'])
-        print("暗杠: ", current_player_meld['an_gang'])
-        print("弃牌: ", current_player_meld['discards'])
+        if current_player_meld['peng']:         print("碰牌: ", current_player_meld['peng'])
+        if current_player_meld['chi']:          print("吃牌: ", current_player_meld['chi'])
+        if current_player_meld['ming_gang']:    print("明杠: ", current_player_meld['ming_gang'])
+        if current_player_meld['an_gang']:      print("暗杠: ", current_player_meld['an_gang'])
+        if current_player_meld['discards']:     print("弃牌: ", current_player_meld['discards'])
         print("--------------------")
         return
 
@@ -93,14 +100,14 @@ class EngineVisualization:
         for c in range(4):
             for v in range(9):
                 if peng[c][v]:
-                    meld['peng'].append(TILE_NAMES[c][v])
+                    meld['peng'].append(self.true_tile_names[c][v])
         # 转换明杠信息
         ming_gang = state['ming_gang'][playerid]
         meld.setdefault('ming_gang', [])
         for c in range(4):
             for v in range(9):
                 if ming_gang[c][v]:
-                    meld['ming_gang'].append(TILE_NAMES[c][v])
+                    meld['ming_gang'].append(self.true_tile_names[c][v])
         # 转换弃牌信息
         discards = state['discards'][playerid][:state['discard_ptr'][playerid]]
         meld.setdefault('discards', [])
@@ -109,7 +116,7 @@ class EngineVisualization:
                 continue # 一般不会遇到-1，保险起见留一下
             c = tile_id // 9
             v = tile_id % 9
-            meld['discards'].append(TILE_NAMES[c][v])
+            meld['discards'].append(self.true_tile_names[c][v])
         # 转换吃牌信息
         meld.setdefault('chi', [])
         chi = state['chi'][playerid]
@@ -117,7 +124,7 @@ class EngineVisualization:
             for v in range(7):
                 count = chi[c][v]
                 for _ in range(count):
-                    meld['chi'].append(f"{TILE_NAMES[c][v]},{TILE_NAMES[c][v+1]},{TILE_NAMES[c][v+2]}") # 吃牌显示为三个牌的组合
+                    meld['chi'].append(f"{self.true_tile_names[c][v]},{self.true_tile_names[c][v+1]},{self.true_tile_names[c][v+2]}") # 吃牌显示为三个牌的组合
         # 转换暗杠数量
         meld['an_gang_count'] = state['an_gang_count'][playerid]
         
@@ -128,21 +135,52 @@ class EngineVisualization:
             for c in range(4):
                 for v in range(9):
                     if an_gang[c][v]:
-                        meld['an_gang'].append(TILE_NAMES[c][v])
+                        meld['an_gang'].append(self.true_tile_names[c][v])
             # 转换自己手牌信息
             meld['hand'] = []
             hand = state['hand']
             for c in range(4):
                 for v in range(9):
                     count = int(hand[c][v])
-                    meld['hand'].extend([TILE_NAMES[c][v]] * count)
+                    meld['hand'].extend([self.true_tile_names[c][v]] * count)
         else:
             meld['an_gang'] = [] # 暗杠信息不显示具体牌面，只显示数量
 
         return meld
     
-    def display_last_action(self, action: int) -> None:
+    def display_last_action(self, last_action: dict[str, int | None]) -> None:
         """显示上一个玩家的动作，包括动作类型和涉及的牌面信息
         Args:
-            action(int): 上一个玩家的动作编码，整数表示不同的动作类型和牌面信息"""
-        pass
+            last_action(dict): 上一个玩家的动作信息，包含动作类型和涉及的牌面信息"""
+        if last_action['playerid'] is None:
+            print("各家摸牌完毕，正在判定三财神、八对...")
+            return # 出现在游戏开始时判定三财神、八对的情况
+        if last_action['playerid'] == -1:
+            print("当前玩家摸牌: ", self.true_tile_names[last_action['action_id']//9][last_action['action_id']%9])
+        else:
+            print(f"刚刚玩家{last_action['playerid']}", end="")
+            if last_action['action_id'] == 0:
+                print("过牌") # 正常不应该触发这个
+            elif last_action['action_id'] == 1:
+                print("胡牌") # 正常不应该触发这个
+            elif 2 <= last_action['action_id'] <= 35:
+                print("弃掉了", self.true_tile_names[(last_action['action_id']-2)//9][(last_action['action_id']-2)%9])
+            elif 36 <= last_action['action_id'] <= 69:
+                print("暗杠") # 不可显示具体牌面信息
+            elif 70 <= last_action['action_id'] <= 103:
+                print("明杠了", self.true_tile_names[(last_action['action_id']-70)//9][(last_action['action_id']-70)%9])
+            elif 104 <= last_action['action_id'] <= 137:
+                print("补杠了", self.true_tile_names[(last_action['action_id']-104)//9][(last_action['action_id']-104)%9]) # 正常不应该触发这个
+            elif 138 <= last_action['action_id'] <= 171:
+                print("碰了", self.true_tile_names[(last_action['action_id']-138)//9][(last_action['action_id']-138)%9])
+            elif 172 <= last_action['action_id'] <= 234:
+                chi_card_type = (last_action['action_id'] - 172) // 21 # 0-2分别对应万条筒
+                chi_card_class = (last_action['action_id'] - 172) % 21 # 对应吃牌的三种组合
+                if chi_card_class <= 6:
+                    print("用", self.true_tile_names[chi_card_type][chi_card_class + 1], "和", self.true_tile_names[chi_card_type][chi_card_class + 2], "吃了", self.true_tile_names[chi_card_type][chi_card_class])
+                elif 7 <= chi_card_class <= 13:
+                    print("用", self.true_tile_names[chi_card_type][chi_card_class - 7], "和", self.true_tile_names[chi_card_type][chi_card_class - 5], "吃了", self.true_tile_names[chi_card_type][chi_card_class - 6])
+                else:
+                    print("用", self.true_tile_names[chi_card_type][chi_card_class - 14], "和", self.true_tile_names[chi_card_type][chi_card_class - 13], "吃了", self.true_tile_names[chi_card_type][chi_card_class - 12])
+        print("--------------------")
+        return